# web/openapi.json
{
  "components": {
    "schemas": {
...
      "ClienteEntrada": {
        "additionalProperties": false,
        "properties": {
          "apellidos": {
            "default": "",
            "maxLength": 200,
            "title": "Apellidos",
            "type": "string"
          },
          "autoriza_tratamiento": {
            "description": "autorización de tratamiento de datos (Ley 1581 de 2012)",
            "title": "Autoriza Tratamiento",
            "type": "boolean"
          },
          "digito_verificacion": {
            "anyOf": [
              {
                "pattern": "^[0-9]$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "solo NIT; si falta se calcula",
            "title": "Digito Verificacion"
          },
          "email": {
            "maxLength": 200,
            "minLength": 5,
            "title": "Email",
            "type": "string"
          },
          "nombres": {
            "description": "para NIT: razón social",
            "maxLength": 200,
            "minLength": 1,
            "title": "Nombres",
            "type": "string"
          },
          "numero_documento": {
            "maxLength": 20,
            "minLength": 3,
            "title": "Numero Documento",
            "type": "string"
          },
          "tipo_documento": {
            "enum": [
              "CC",
              "NIT",
              "NIT_EXTERIOR",
              "CE",
              "PA",
              "PPT"
            ],
            "title": "Tipo Documento",
            "type": "string"
          }
        },
        "required": [
          "tipo_documento",
          "numero_documento",
          "nombres",
          "email",
          "autoriza_tratamiento"
        ],
        "title": "ClienteEntrada",
        "type": "object"
      },
...
      "ClienteSalida": {
        "properties": {
          "activo": {
            "title": "Activo",
            "type": "boolean"
          },
          "apellidos": {
            "title": "Apellidos",
            "type": "string"
          },
          "digito_verificacion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Digito Verificacion"
          },
          "email_enmascarado": {
            "title": "Email Enmascarado",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "nombre_mostrar": {
            "description": "razón social si es NIT; nombres y apellidos si es persona",
            "title": "Nombre Mostrar",
            "type": "string"
          },
          "nombres": {
            "title": "Nombres",
            "type": "string"
          },
          "numero_documento": {
            "title": "Numero Documento",
            "type": "string"
          },
          "tipo_documento": {
            "enum": [
              "CC",
              "NIT",
              "NIT_EXTERIOR",
              "CE",
              "PA",
              "PPT"
            ],
            "title": "Tipo Documento",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tipo_documento",
          "numero_documento",
          "digito_verificacion",
          "nombre_mostrar",
          "nombres",
          "apellidos",
          "email_enmascarado",
          "activo"
        ],
        "title": "ClienteSalida",
        "type": "object"
      },
      "ClientesPaginaSalida": {
        "properties": {
          "filas": {
            "items": {
              "$ref": "#/components/schemas/ClienteSalida"
            },
            "title": "Filas",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "filas",
          "total"
        ],
        "title": "ClientesPaginaSalida",
        "type": "object"
      },
...
      "Error": {
        "properties": {
          "detail": {
            "title": "Detail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "Error",
        "type": "object"
      },
...
      "FacturaEntrada": {
        "additionalProperties": false,
        "properties": {
          "clave": {
            "description": "generada UNA vez por intento de emisión (p. ej. uuid sin guiones); reintentar con la misma clave no duplica la factura",
            "pattern": "^[A-Za-z0-9]{16,64}$",
            "title": "Clave",
            "type": "string"
          },
          "cliente_id": {
            "title": "Cliente Id",
            "type": "integer"
          },
          "medio_pago": {
            "enum": [
              "EFECTIVO",
              "TARJETA"
            ],
            "title": "Medio Pago",
            "type": "string"
          },
          "origen_id": {
            "title": "Origen Id",
            "type": "integer"
          },
          "origen_tipo": {
            "enum": [
              "despacho",
              "venta_manual"
            ],
            "title": "Origen Tipo",
            "type": "string"
          }
        },
        "required": [
          "clave",
          "origen_tipo",
          "origen_id",
          "cliente_id",
          "medio_pago"
        ],
        "title": "FacturaEntrada",
        "type": "object"
      },
      "FacturaSalida": {
        "properties": {
          "cliente": {
            "$ref": "#/components/schemas/ClienteFactura"
          },
          "cufe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cufe"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "estado": {
            "enum": [
              "EN_PROCESO",
              "FACTURADO",
              "RECHAZADO",
              "INCIERTO"
            ],
            "title": "Estado",
            "type": "string"
          },
          "fecha_emision": {
            "format": "date-time",
            "title": "Fecha Emision",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "medio_pago": {
            "enum": [
              "EFECTIVO",
              "TARJETA"
            ],
            "title": "Medio Pago",
            "type": "string"
          },
          "nueva": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "solo al emitir: False si la clave ya tenía factura",
            "title": "Nueva"
          },
          "numero": {
            "title": "Numero",
            "type": "string"
          },
          "origen_id": {
            "title": "Origen Id",
            "type": "integer"
          },
          "origen_tipo": {
            "enum": [
              "despacho",
              "venta_manual"
            ],
            "title": "Origen Tipo",
            "type": "string"
          },
          "total": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total",
            "type": "string"
          }
        },
        "required": [
          "id",
          "numero",
          "estado",
          "total",
          "cufe",
          "fecha_emision",
          "medio_pago",
          "origen_tipo",
          "origen_id",
          "cliente",
          "error"
        ],
        "title": "FacturaSalida",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RegistroPublicoEntrada": {
        "additionalProperties": false,
        "properties": {
          "apellidos": {
            "default": "",
            "maxLength": 200,
            "title": "Apellidos",
            "type": "string"
          },
          "autoriza_tratamiento": {
            "description": "autorización de tratamiento de datos (Ley 1581 de 2012)",
            "title": "Autoriza Tratamiento",
            "type": "boolean"
          },
          "captcha": {
            "anyOf": [
              {
                "maxLength": 4096,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "token del widget; obligatorio si la API tiene captcha configurado",
            "title": "Captcha"
          },
          "digito_verificacion": {
            "anyOf": [
              {
                "pattern": "^[0-9]$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "solo NIT; si falta se calcula",
            "title": "Digito Verificacion"
          },
          "email": {
            "maxLength": 200,
            "minLength": 5,
            "title": "Email",
            "type": "string"
          },
          "nombres": {
            "description": "para NIT: razón social",
            "maxLength": 200,
            "minLength": 1,
            "title": "Nombres",
            "type": "string"
          },
          "numero_documento": {
            "maxLength": 20,
            "minLength": 3,
            "title": "Numero Documento",
            "type": "string"
          },
          "tipo_documento": {
            "enum": [
              "CC",
              "NIT",
              "NIT_EXTERIOR",
              "CE",
              "PA",
              "PPT"
            ],
            "title": "Tipo Documento",
            "type": "string"
          }
        },
        "required": [
          "tipo_documento",
          "numero_documento",
          "nombres",
          "email",
          "autoriza_tratamiento"
        ],
        "title": "RegistroPublicoEntrada",
        "type": "object"
      },
      "RegistroPublicoSalida": {
        "properties": {
          "mensaje": {
            "title": "Mensaje",
            "type": "string"
          }
        },
        "required": [
          "mensaje"
        ],
        "title": "RegistroPublicoSalida",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "token de POST /auth/login",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API del sistema que reemplaza a Nexus. **Banco de pruebas: no envía nada a la DIAN.** Dinero y galones viajan como texto decimal.",
    "title": "Estación de servicio — banco de pruebas",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/clientes": {
      "get": {
        "operationId": "listar_clientes_clientes_get",
        "parameters": [
          {
            "description": "inicio del documento o parte del nombre",
            "in": "query",
            "name": "texto",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "inicio del documento o parte del nombre",
              "title": "Texto"
            }
          },
          {
            "in": "query",
            "name": "estado",
            "required": false,
            "schema": {
              "default": "activos",
              "enum": [
                "activos",
                "inactivos",
                "todos"
              ],
              "title": "Estado",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limite",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limite",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "desplazamiento",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Desplazamiento",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientesPaginaSalida"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Listar y buscar clientes",
        "tags": [
          "administración"
        ]
      },
      "post": {
        "operationId": "crear_clientes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClienteEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteSalida"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Registrar un cliente desde la pista (exige la autorización de datos)",
        "tags": [
          "clientes"
        ]
      }
    },
    "/clientes/buscar": {
      "get": {
        "operationId": "buscar_clientes_buscar_get",
        "parameters": [
          {
            "in": "query",
            "name": "tipo",
            "required": true,
            "schema": {
              "enum": [
                "CC",
                "NIT",
                "NIT_EXTERIOR",
                "CE",
                "PA",
                "PPT"
              ],
              "title": "Tipo",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "numero",
            "required": true,
            "schema": {
              "maxLength": 20,
              "minLength": 3,
              "title": "Numero",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteSalida"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Buscar un cliente por tipo y número de documento",
        "tags": [
          "clientes"
        ]
      }
    },
...
    "/facturas": {
      "post": {
        "operationId": "emitir_facturas_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FacturaEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacturaSalida"
                }
              }
            },
            "description": "la clave ya tenía factura (doble clic o reintento)"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacturaSalida"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "la venta ya no está pendiente o la clave se usó para otra venta"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "datos inválidos o caso todavía no soportado"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "falta configuración fiscal o numeración: no se emite"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Emitir la factura de un despacho o venta manual (idempotente por `clave`)",
        "tags": [
          "facturas"
        ]
      }
    },
    "/facturas/{factura_id}": {
      "get": {
        "operationId": "consultar_facturas__factura_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "factura_id",
            "required": true,
            "schema": {
              "title": "Factura Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacturaSalida"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Estado de una factura",
        "tags": [
          "facturas"
        ]
      }
    },
    "/facturas/{factura_id}/pdf": {
      "get": {
        "operationId": "pdf_facturas__factura_id__pdf_get",
        "parameters": [
          {
            "in": "path",
            "name": "factura_id",
            "required": true,
            "schema": {
              "title": "Factura Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/pdf": {}
            },
            "description": "representación gráfica (PDF)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "la factura todavía no está validada"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "el XML no permite armar la representación"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Representación gráfica (PDF) de una factura validada, armada desde su XML",
        "tags": [
          "facturas"
        ]
      }
    },
...
    "/publico/registro": {
      "post": {
        "description": "Respuesta IGUAL si el documento ya estaba registrado: así no se puede averiguar quién es cliente.\nUn registro público nunca modifica los datos de un cliente existente.",
        "operationId": "registro_publico_publico_registro_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistroPublicoEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistroPublicoSalida"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Too Many Requests"
          }
        },
        "summary": "Autorregistro del cliente por QR (público, con límite de tasa)",
        "tags": [
          "clientes",
          "público"
        ]
      }
    },
...
  }
}